Use floats for adding scores, the same way as in the WordPruningBreadthFirstSearchManager, to avoid round off differences.
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@3730 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ edu/cmu/sphinx/result/TokenGraphToLattice.java
@@ -126,7 +126,7 @@ public class TokenGraphToLattice {
      *                      parent of token
      */
     private void collapseWordPath(Node parentWordNode, Token token,
-                                  double acousticScore, double languageScore) {
+                                  float acousticScore, float languageScore) {
         if (token.isWord()) {
             /*
              * If this is a word, create a Node for it, and then create an
@@ -134,7 +134,7 @@ public class TokenGraphToLattice {
              */
             Node fromNode = getNode(token);
             edges.add(new Edge(fromNode, parentWordNode,
-                               acousticScore, languageScore));
+                               (double)acousticScore, (double)languageScore));
 
             if (token.getPredecessor() != null) {
                 /* Collapse the token sequence ending in this token. */
